Rename join -> joins package.
