+++ sql/core/src/main/scala/org/apache/spark/sql/execution/joins/CartesianProduct.scala
@@ -15,23 +15,26 @@
  * limitations under the License.
  */
 
-package org.apache.spark.sql.execution
+package org.apache.spark.sql.execution.joins
 
 import org.apache.spark.annotation.DeveloperApi
+import org.apache.spark.sql.catalyst.expressions.JoinedRow
+import org.apache.spark.sql.execution.{BinaryNode, SparkPlan}
 
 /**
  * :: DeveloperApi ::
- * Physical execution operators for join operations.
  */
-package object join {
+@DeveloperApi
+case class CartesianProduct(left: SparkPlan, right: SparkPlan) extends BinaryNode {
+  override def output = left.output ++ right.output
 
-  @DeveloperApi
-  sealed abstract class BuildSide
+  override def execute() = {
+    val leftResults = left.execute().map(_.copy())
+    val rightResults = right.execute().map(_.copy())
 
-  @DeveloperApi
-  case object BuildRight extends BuildSide
-
-  @DeveloperApi
-  case object BuildLeft extends BuildSide
-
+    leftResults.cartesian(rightResults).mapPartitions { iter =>
+      val joinedRow = new JoinedRow
+      iter.map(r => joinedRow(r._1, r._2))
+    }
+  }
 }
